Use $XDG_DATA_DIR as default history path
Fixes #10100

Consensus was that the history file was not config and so should not live
in the config directory by default.

Also includes an automated "migration" that moves the old history file
to the new path: only if the old file exists and there is no file in the
new path.

Notes:
  * Changes `nu_path::data_dir()` to `nu_path::nu_data_dir()` so that it
    returns the data dir with `/nushell` appended. This was already
    being doing by callers of `nu_path::data_dir()` so I just refactored
    it.
  * `history_import.rs` tests now set `XDG_CONFIG_HOME` _and_
    `XDG_DATA_HOME`, there may be other tests that will come to need
    both. But currently all tests pass.