Update error message test for split-by
The now deprecated split-by command produces different error messages
depending on whether the input is a value or a stream.

The `errors_if_non_record_input` test used to check whether the error
message produced was the one produced when the input is a value. This
commit updates this test so that is also accepts the error message
produced when the input is a stream.